refactor: decompose state-machine.ts into guards.ts and hsm-definitions.ts
Extract guard definitions (26 guards, Guard/GuardResult types, hasArtifact,
collectReviewStatuses, PASSED/FAILED_STATUSES) into guards.ts (343L).

Extract HSM factory functions (createFeatureHSM, createDebugHSM,
createRefactorHSM) into hsm-definitions.ts (270L).

state-machine.ts reduced from 1008 to 406 lines. Re-exports Guard/GuardResult
types for backward compatibility.
