+++ plugins/exarchos/servers/exarchos-mcp/src/workflow/state-machine.ts
@@ -1,14 +1,12 @@
+import type { Guard, GuardResult } from './guards.js';
+import { createFeatureHSM, createDebugHSM, createRefactorHSM } from './hsm-definitions.js';
+
+// Re-export guard types for consumers
+export type { Guard, GuardResult };
+
 // ─── HSM Types ──────────────────────────────────────────────────────────────
 
 export type Effect = 'checkpoint' | 'log' | 'increment-fix-cycle';
-
-export type GuardResult = boolean | { readonly passed: false; readonly reason: string };
-
-export interface Guard {
-  readonly id: string;
-  readonly evaluate: (state: Record<string, unknown>) => GuardResult;
-  readonly description: string;
-}
 
 export interface State {
   readonly id: string;
@@ -55,606 +53,6 @@ export interface TransitionResult {
   readonly errorMessage?: string;
   readonly guardDescription?: string;
   readonly validTargets?: readonly string[];
-}
-
-// ─── Guards ─────────────────────────────────────────────────────────────────
-
-function hasArtifact(field: string): Guard {
-  return {
-    id: `${field}-exists`,
-    description: `${field} artifact must exist`,
-    evaluate: (state: Record<string, unknown>) => {
-      const artifacts = state.artifacts as Record<string, unknown> | undefined;
-      return artifacts != null && artifacts[field] != null;
-    },
-  };
-}
-
-const PASSED_STATUSES = new Set(['pass', 'passed', 'approved']);
-const FAILED_STATUSES = new Set(['fail', 'failed', 'needs_fixes']);
-
-/**
- * Collects all `status` field values from a reviews object, handling both flat
- * and nested shapes:
- *   - flat:   reviews.overhaul = { status: "approved", ... }
- *   - nested: reviews.A1 = { specReview: { status: "pass" }, qualityReview: { status: "approved" } }
- * Also supports the legacy `passed: boolean` shape for backward compatibility.
- */
-function collectReviewStatuses(
-  reviews: Record<string, unknown>,
-): Array<{ path: string; status: string }> {
-  const results: Array<{ path: string; status: string }> = [];
-  for (const [key, value] of Object.entries(reviews)) {
-    if (typeof value !== 'object' || value === null) continue;
-    const entry = value as Record<string, unknown>;
-    if (typeof entry.status === 'string') {
-      // Flat review: { status: "approved", ... }
-      results.push({ path: key, status: entry.status });
-    } else if (typeof entry.passed === 'boolean') {
-      // Legacy: { passed: true/false }
-      results.push({ path: key, status: entry.passed ? 'passed' : 'failed' });
-    } else {
-      // Nested: { specReview: { status: "pass" }, qualityReview: { status: "approved" } }
-      for (const [subKey, subValue] of Object.entries(entry)) {
-        if (typeof subValue !== 'object' || subValue === null) continue;
-        const sub = subValue as Record<string, unknown>;
-        if (typeof sub.status === 'string') {
-          results.push({ path: `${key}.${subKey}`, status: sub.status });
-        } else if (typeof sub.passed === 'boolean') {
-          results.push({ path: `${key}.${subKey}`, status: sub.passed ? 'passed' : 'failed' });
-        }
-      }
-    }
-  }
-  return results;
-}
-
-const guards = {
-  designArtifactExists: {
-    id: 'design-artifact-exists',
-    description: 'Design artifact must exist',
-    evaluate: (state: Record<string, unknown>) => {
-      const artifacts = state.artifacts as Record<string, unknown> | undefined;
-      return artifacts != null && artifacts.design != null;
-    },
-  },
-
-  planArtifactExists: {
-    id: 'plan-artifact-exists',
-    description: 'Plan artifact must exist',
-    evaluate: (state: Record<string, unknown>) => {
-      const artifacts = state.artifacts as Record<string, unknown> | undefined;
-      return artifacts != null && artifacts.plan != null;
-    },
-  },
-
-  allTasksComplete: {
-    id: 'all-tasks-complete',
-    description: 'All tasks must be complete',
-    evaluate: (state: Record<string, unknown>) => {
-      const tasks = state.tasks as Array<{ status: string }> | undefined;
-      if (!tasks || tasks.length === 0) return true;
-      return tasks.every((t) => t.status === 'complete');
-    },
-  },
-
-  allReviewsPassed: {
-    id: 'all-reviews-passed',
-    description: 'All reviews must have passed',
-    evaluate: (state: Record<string, unknown>): GuardResult => {
-      const reviews = state.reviews as Record<string, unknown> | undefined;
-      if (!reviews) {
-        return {
-          passed: false,
-          reason:
-            'state.reviews is missing — set reviews.{name} with status: "pass" or "approved"',
-        };
-      }
-      const statuses = collectReviewStatuses(reviews);
-      if (statuses.length === 0) {
-        return {
-          passed: false,
-          reason:
-            'state.reviews has no recognizable review entries — each review needs a status field ("pass", "approved", "fail", "needs_fixes")',
-        };
-      }
-      const notPassed = statuses.filter((s) => !PASSED_STATUSES.has(s.status));
-      if (notPassed.length > 0) {
-        return {
-          passed: false,
-          reason: `Reviews not passed: ${notPassed.map((s) => `${s.path} (status: "${s.status}")`).join(', ')}`,
-        };
-      }
-      return true;
-    },
-  },
-
-  anyReviewFailed: {
-    id: 'any-review-failed',
-    description: 'At least one review must have failed',
-    evaluate: (state: Record<string, unknown>): GuardResult => {
-      const reviews = state.reviews as Record<string, unknown> | undefined;
-      if (!reviews) {
-        return {
-          passed: false,
-          reason: 'state.reviews is missing — cannot determine if any review failed',
-        };
-      }
-      const statuses = collectReviewStatuses(reviews);
-      if (statuses.length === 0) {
-        return { passed: false, reason: 'state.reviews has no recognizable review entries' };
-      }
-      const hasFailed = statuses.some((s) => FAILED_STATUSES.has(s.status));
-      if (!hasFailed) {
-        return {
-          passed: false,
-          reason: `No failed reviews found: ${statuses.map((s) => `${s.path} (status: "${s.status}")`).join(', ')}`,
-        };
-      }
-      return true;
-    },
-  },
-
-  prUrlExists: {
-    id: 'pr-url-exists',
-    description: 'PR URL must exist',
-    evaluate: (state: Record<string, unknown>) => {
-      const synthesis = state.synthesis as Record<string, unknown> | undefined;
-      if (synthesis?.prUrl != null) return true;
-      const artifacts = state.artifacts as Record<string, unknown> | undefined;
-      return artifacts?.pr != null;
-    },
-  },
-
-  humanUnblocked: {
-    id: 'human-unblocked',
-    description: 'Human must have unblocked the workflow',
-    evaluate: (state: Record<string, unknown>) => {
-      return state.unblocked === true;
-    },
-  },
-
-  triageComplete: {
-    id: 'triage-complete',
-    description: 'Triage must be complete',
-    evaluate: (state: Record<string, unknown>) => {
-      const triage = state.triage as Record<string, unknown> | undefined;
-      return triage != null && triage.symptom != null;
-    },
-  },
-
-  rootCauseFound: {
-    id: 'root-cause-found',
-    description: 'Root cause must be identified',
-    evaluate: (state: Record<string, unknown>) => {
-      const investigation = state.investigation as Record<string, unknown> | undefined;
-      return investigation != null && investigation.rootCause != null;
-    },
-  },
-
-  hotfixTrackSelected: {
-    id: 'hotfix-track-selected',
-    description: 'Hotfix track must be selected',
-    evaluate: (state: Record<string, unknown>) => {
-      return state.track === 'hotfix';
-    },
-  },
-
-  thoroughTrackSelected: {
-    id: 'thorough-track-selected',
-    description: 'Thorough track must be selected',
-    evaluate: (state: Record<string, unknown>) => {
-      return state.track === 'thorough';
-    },
-  },
-
-  rcaDocumentComplete: {
-    id: 'rca-document-complete',
-    description: 'RCA document must be complete',
-    evaluate: (state: Record<string, unknown>) => {
-      const artifacts = state.artifacts as Record<string, unknown> | undefined;
-      return artifacts?.rca != null;
-    },
-  },
-
-  fixDesignComplete: {
-    id: 'fix-design-complete',
-    description: 'Fix design must be complete',
-    evaluate: (state: Record<string, unknown>) => {
-      const artifacts = state.artifacts as Record<string, unknown> | undefined;
-      return artifacts?.fixDesign != null;
-    },
-  },
-
-  implementationComplete: {
-    id: 'implementation-complete',
-    description: 'Implementation must be complete',
-    evaluate: () => true,
-  },
-
-  validationPassed: {
-    id: 'validation-passed',
-    description: 'Validation must have passed',
-    evaluate: (state: Record<string, unknown>) => {
-      const validation = state.validation as Record<string, unknown> | undefined;
-      return validation != null && validation.testsPass === true;
-    },
-  },
-
-  reviewPassed: {
-    id: 'review-passed',
-    description: 'Review must have passed',
-    evaluate: (state: Record<string, unknown>): GuardResult => {
-      const reviews = state.reviews as Record<string, unknown> | undefined;
-      if (!reviews) {
-        return {
-          passed: false,
-          reason:
-            'state.reviews is missing — set reviews.{name} with status: "pass" or "approved"',
-        };
-      }
-      const statuses = collectReviewStatuses(reviews);
-      if (statuses.length === 0) {
-        return {
-          passed: false,
-          reason:
-            'state.reviews has no recognizable review entries — each review needs a status field',
-        };
-      }
-      const notPassed = statuses.filter((s) => !PASSED_STATUSES.has(s.status));
-      if (notPassed.length > 0) {
-        return {
-          passed: false,
-          reason: `Reviews not passed: ${notPassed.map((s) => `${s.path} (status: "${s.status}")`).join(', ')}`,
-        };
-      }
-      return true;
-    },
-  },
-
-  scopeAssessmentComplete: {
-    id: 'scope-assessment-complete',
-    description: 'Scope assessment must be complete',
-    evaluate: (state: Record<string, unknown>) => {
-      const explore = state.explore as Record<string, unknown> | undefined;
-      return explore?.scopeAssessment != null;
-    },
-  },
-
-  briefComplete: {
-    id: 'brief-complete',
-    description: 'Brief must be complete',
-    evaluate: (state: Record<string, unknown>) => {
-      const brief = state.brief as Record<string, unknown> | undefined;
-      return brief != null && brief.goals != null;
-    },
-  },
-
-  polishTrackSelected: {
-    id: 'polish-track-selected',
-    description: 'Polish track must be selected',
-    evaluate: (state: Record<string, unknown>) => {
-      return state.track === 'polish';
-    },
-  },
-
-  overhaulTrackSelected: {
-    id: 'overhaul-track-selected',
-    description: 'Overhaul track must be selected',
-    evaluate: (state: Record<string, unknown>) => {
-      return state.track === 'overhaul';
-    },
-  },
-
-  docsUpdated: {
-    id: 'docs-updated',
-    description: 'Documentation must be updated',
-    evaluate: (state: Record<string, unknown>) => {
-      const validation = state.validation as Record<string, unknown> | undefined;
-      return validation?.docsUpdated === true;
-    },
-  },
-
-  goalsVerified: {
-    id: 'goals-verified',
-    description: 'Refactor goals must be verified',
-    evaluate: (state: Record<string, unknown>) => {
-      const validation = state.validation as Record<string, unknown> | undefined;
-      return validation?.testsPass === true;
-    },
-  },
-
-  planReviewComplete: {
-    id: 'plan-review-complete',
-    description: 'Plan review must be complete with no gaps',
-    evaluate: (state: Record<string, unknown>) => {
-      const planReview = state.planReview as Record<string, unknown> | undefined;
-      return planReview?.approved === true;
-    },
-  },
-
-  planReviewGapsFound: {
-    id: 'plan-review-gaps-found',
-    description: 'Plan review found coverage gaps',
-    evaluate: (state: Record<string, unknown>) => {
-      const planReview = state.planReview as Record<string, unknown> | undefined;
-      return planReview?.gapsFound === true;
-    },
-  },
-
-  always: {
-    id: 'always',
-    description: 'Always passes',
-    evaluate: () => true,
-  },
-} as const satisfies Record<string, Guard>;
-
-// ─── Feature Workflow HSM ───────────────────────────────────────────────────
-
-function createFeatureHSM(): HSMDefinition {
-  const states: Record<string, State> = {
-    ideate: { id: 'ideate', type: 'atomic' },
-    plan: { id: 'plan', type: 'atomic' },
-    'plan-review': { id: 'plan-review', type: 'atomic' },
-    implementation: {
-      id: 'implementation',
-      type: 'compound',
-      initial: 'delegate',
-      maxFixCycles: 3,
-      onEntry: ['log'],
-      onExit: ['log'],
-    },
-    delegate: { id: 'delegate', type: 'atomic', parent: 'implementation' },
-    review: { id: 'review', type: 'atomic', parent: 'implementation' },
-    synthesize: { id: 'synthesize', type: 'atomic' },
-    completed: { id: 'completed', type: 'final' },
-    cancelled: { id: 'cancelled', type: 'final' },
-    blocked: { id: 'blocked', type: 'atomic' },
-  };
-
-  const transitions: Transition[] = [
-    { from: 'ideate', to: 'plan', guard: guards.designArtifactExists },
-    { from: 'plan', to: 'plan-review', guard: guards.planArtifactExists },
-    { from: 'plan-review', to: 'delegate', guard: guards.planReviewComplete },
-    {
-      from: 'plan-review',
-      to: 'plan',
-      guard: guards.planReviewGapsFound,
-      effects: ['log'],
-    },
-    { from: 'delegate', to: 'review', guard: guards.allTasksComplete },
-    { from: 'review', to: 'synthesize', guard: guards.allReviewsPassed },
-    {
-      from: 'review',
-      to: 'delegate',
-      guard: guards.anyReviewFailed,
-      isFixCycle: true,
-      effects: ['increment-fix-cycle'],
-    },
-    { from: 'synthesize', to: 'completed', guard: guards.prUrlExists },
-    { from: 'blocked', to: 'delegate', guard: guards.humanUnblocked },
-  ];
-
-  return { id: 'feature', states, transitions };
-}
-
-// ─── Debug Workflow HSM ─────────────────────────────────────────────────────
-
-function createDebugHSM(): HSMDefinition {
-  const states: Record<string, State> = {
-    triage: { id: 'triage', type: 'atomic' },
-    investigate: { id: 'investigate', type: 'atomic' },
-
-    // Thorough track compound
-    'thorough-track': {
-      id: 'thorough-track',
-      type: 'compound',
-      initial: 'rca',
-      maxFixCycles: 2,
-      onEntry: ['log'],
-      onExit: ['log'],
-    },
-    rca: { id: 'rca', type: 'atomic', parent: 'thorough-track' },
-    design: { id: 'design', type: 'atomic', parent: 'thorough-track' },
-    'debug-implement': {
-      id: 'debug-implement',
-      type: 'atomic',
-      parent: 'thorough-track',
-    },
-    'debug-validate': {
-      id: 'debug-validate',
-      type: 'atomic',
-      parent: 'thorough-track',
-    },
-    'debug-review': {
-      id: 'debug-review',
-      type: 'atomic',
-      parent: 'thorough-track',
-    },
-
-    // Hotfix track compound
-    'hotfix-track': {
-      id: 'hotfix-track',
-      type: 'compound',
-      initial: 'hotfix-implement',
-      onEntry: ['log'],
-      onExit: ['log'],
-    },
-    'hotfix-implement': {
-      id: 'hotfix-implement',
-      type: 'atomic',
-      parent: 'hotfix-track',
-    },
-    'hotfix-validate': {
-      id: 'hotfix-validate',
-      type: 'atomic',
-      parent: 'hotfix-track',
-    },
-
-    synthesize: { id: 'synthesize', type: 'atomic' },
-    completed: { id: 'completed', type: 'final' },
-    cancelled: { id: 'cancelled', type: 'final' },
-    blocked: { id: 'blocked', type: 'atomic' },
-  };
-
-  const transitions: Transition[] = [
-    { from: 'triage', to: 'investigate', guard: guards.triageComplete },
-
-    // Investigate → thorough or hotfix
-    { from: 'investigate', to: 'rca', guard: guards.thoroughTrackSelected },
-    {
-      from: 'investigate',
-      to: 'hotfix-implement',
-      guard: guards.hotfixTrackSelected,
-    },
-
-    // Thorough track flow
-    { from: 'rca', to: 'design', guard: guards.rcaDocumentComplete },
-    { from: 'design', to: 'debug-implement', guard: guards.fixDesignComplete },
-    {
-      from: 'debug-implement',
-      to: 'debug-validate',
-      guard: guards.implementationComplete,
-    },
-    { from: 'debug-validate', to: 'debug-review', guard: guards.validationPassed },
-    { from: 'debug-review', to: 'synthesize', guard: guards.reviewPassed },
-
-    // Hotfix track flow
-    {
-      from: 'hotfix-implement',
-      to: 'hotfix-validate',
-      guard: guards.implementationComplete,
-    },
-    { from: 'hotfix-validate', to: 'completed', guard: guards.validationPassed },
-
-    // Synthesize → completed
-    { from: 'synthesize', to: 'completed', guard: guards.prUrlExists },
-  ];
-
-  return { id: 'debug', states, transitions };
-}
-
-// ─── Refactor Workflow HSM ──────────────────────────────────────────────────
-
-function createRefactorHSM(): HSMDefinition {
-  const states: Record<string, State> = {
-    explore: { id: 'explore', type: 'atomic' },
-    brief: { id: 'brief', type: 'atomic' },
-
-    // Polish track compound
-    'polish-track': {
-      id: 'polish-track',
-      type: 'compound',
-      initial: 'polish-implement',
-      onEntry: ['log'],
-      onExit: ['log'],
-    },
-    'polish-implement': {
-      id: 'polish-implement',
-      type: 'atomic',
-      parent: 'polish-track',
-    },
-    'polish-validate': {
-      id: 'polish-validate',
-      type: 'atomic',
-      parent: 'polish-track',
-    },
-    'polish-update-docs': {
-      id: 'polish-update-docs',
-      type: 'atomic',
-      parent: 'polish-track',
-    },
-
-    // Overhaul track compound
-    'overhaul-track': {
-      id: 'overhaul-track',
-      type: 'compound',
-      initial: 'overhaul-plan',
-      maxFixCycles: 3,
-      onEntry: ['log'],
-      onExit: ['log'],
-    },
-    'overhaul-plan': {
-      id: 'overhaul-plan',
-      type: 'atomic',
-      parent: 'overhaul-track',
-    },
-    'overhaul-delegate': {
-      id: 'overhaul-delegate',
-      type: 'atomic',
-      parent: 'overhaul-track',
-    },
-    'overhaul-review': {
-      id: 'overhaul-review',
-      type: 'atomic',
-      parent: 'overhaul-track',
-    },
-    'overhaul-update-docs': {
-      id: 'overhaul-update-docs',
-      type: 'atomic',
-      parent: 'overhaul-track',
-    },
-
-    synthesize: { id: 'synthesize', type: 'atomic' },
-    completed: { id: 'completed', type: 'final' },
-    cancelled: { id: 'cancelled', type: 'final' },
-    blocked: { id: 'blocked', type: 'atomic' },
-  };
-
-  const transitions: Transition[] = [
-    { from: 'explore', to: 'brief', guard: guards.scopeAssessmentComplete },
-
-    // Brief → polish or overhaul
-    {
-      from: 'brief',
-      to: 'polish-implement',
-      guard: guards.polishTrackSelected,
-    },
-    { from: 'brief', to: 'overhaul-plan', guard: guards.overhaulTrackSelected },
-
-    // Polish track flow
-    {
-      from: 'polish-implement',
-      to: 'polish-validate',
-      guard: guards.implementationComplete,
-    },
-    {
-      from: 'polish-validate',
-      to: 'polish-update-docs',
-      guard: guards.goalsVerified,
-    },
-    { from: 'polish-update-docs', to: 'completed', guard: guards.docsUpdated },
-
-    // Overhaul track flow
-    {
-      from: 'overhaul-plan',
-      to: 'overhaul-delegate',
-      guard: guards.planArtifactExists,
-    },
-    {
-      from: 'overhaul-delegate',
-      to: 'overhaul-review',
-      guard: guards.allTasksComplete,
-    },
-    {
-      from: 'overhaul-review',
-      to: 'overhaul-update-docs',
-      guard: guards.allReviewsPassed,
-    },
-    {
-      from: 'overhaul-review',
-      to: 'overhaul-delegate',
-      guard: guards.anyReviewFailed,
-      isFixCycle: true,
-      effects: ['increment-fix-cycle'],
-    },
-    { from: 'overhaul-update-docs', to: 'synthesize', guard: guards.docsUpdated },
-
-    // Synthesize → completed
-    { from: 'synthesize', to: 'completed', guard: guards.prUrlExists },
-  ];
-
-  return { id: 'refactor', states, transitions };
 }
 
 // ─── HSM Registry ───────────────────────────────────────────────────────────
